refactor: util 함수 폴더 구조 수정
