--- src/utils/apiClient.js
@@ -1,9 +1,0 @@
-import axios from 'axios';
-const { REACT_APP_API_URL } = process.env;
-
-const apiClient = axios.create({
-  baseURL: REACT_APP_API_URL,
-  responseType: 'json',
-});
-
-export default apiClient;
