--- src/utils/parsePrice.js
@@ -1,5 +1,0 @@
-const parsePrice = (price) => {
-  return price.toLocaleString('ko-KR');
-};
-
-export default parsePrice;
